fix: 将 protocolVersion 类型从 byte 改为 short

+++ debug/xtream-codec-server-reactive-debug-tcp/src/main/java/io/github/hylexus/xtream/debug/codec/server/reactive/tcp/handler/DemoTcpXtreamHandler2.java
@@ -71,7 +71,7 @@ public class DemoTcpXtreamHandler2 {
         final DemoLocationMsg01 msg = new DemoLocationMsg01();
         msg.setMsgId(0x01);
         msg.setMsgBodyProps(222);
-        msg.setProtocolVersion((byte) 1);
+        msg.setProtocolVersion((short) 1);
         msg.setTerminalId("6666");
         // return Flux.error(new RuntimeException("..."));
         // return ByteBufAllocator.DEFAULT.buffer().writeBytes(new byte[]{1,1,2,2});
